refactor(ui): replace multi-step flow with simplified RootView
Consolidate app entry point to a single RootView that handles image
selection and directly presents TextifyView. Remove legacy screens
(Home, ImageSelection, Generation, TextInput, Result, Main) that are
no longer needed with the streamlined UX flow.
